refactor web server to be usable from other commands
