+++ src/Microsoft.DotNet.XHarness.CLI/CommandArguments/TestCommandArguments.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Collections.Generic;
+using System.IO;
 using Mono.Options;
 
 namespace Microsoft.DotNet.XHarness.CLI.CommandArguments
@@ -21,6 +23,8 @@ namespace Microsoft.DotNet.XHarness.CLI.CommandArguments
         /// Tests classes to be included in the run while all others are ignored.
         /// </summary>
         public IEnumerable<string> ClassMethodFilters => _classMethodFilters;
+        public IList<string> WebServerMiddlewarePaths { get; set; } = new List<string>();
+        public bool SetWebServerEnvironmentVariables { get; set; } = false;
 
         protected override OptionSet GetCommandOptions()
         {
@@ -39,6 +43,19 @@ namespace Microsoft.DotNet.XHarness.CLI.CommandArguments
                     "tests that have been provided by the '--method' and '--class' arguments will be ran. All other test will be " +
                     "ignored. Can be used more than once.",
                     v => _classMethodFilters.Add(v)
+                },
+                { "web-server-middleware=", "Path to assembly which contains middleware for endpoints for local test server.",
+                    v =>
+                    {
+                        if (!File.Exists(v))
+                        {
+                            throw new ArgumentException($"Failed to find the middleware assembly at {v}");
+                        }
+                        WebServerMiddlewarePaths.Add(v);
+                    }
+                },
+                { "set-web-server-env", "Set environment variables, so that unit test use xharness as test web server.",
+                    v => SetWebServerEnvironmentVariables = true
                 },
             };
